lambda/dev/requirements.txt
moviepy==1.0.3
Pillow==10.1.0
boto3>=1.26.0
botocore>=1.29.0
opencv-python>=4.8.0
mediapipe>=0.10.0
ffmpeg-python>=0.2.0
firebase-admin>=6.0.0


